move the changes to indexing/insert.py
