+++ tests/cupy_tests/manipulation_tests/test_basic.py
@@ -126,81 +126,3 @@ class TestCopytoFromScalar(unittest.TestCase):
             self.dst_shape, xp, dtype) % 2).astype(xp.bool_)
         xp.copyto(dst, self.src, where=mask)
         return dst
-
-
-@testing.parameterize(
-    *testing.product(
-        {'shape': [(0,), (1,), (2, 3), (2, 3, 4)]}))
-@testing.gpu
-class TestPutmaskEqual(unittest.TestCase):
-
-    @testing.for_all_dtypes()
-    @testing.numpy_cupy_allclose()
-    def test_putmask(self, xp, dtype):
-        a = testing.shaped_random(self.shape, xp, dtype=dtype, seed=0)
-        mask = testing.shaped_random(self.shape, xp, dtype=bool, seed=1)
-        values = testing.shaped_random(self.shape, xp, dtype=dtype, seed=2)
-        ret = xp.putmask(a, mask, values)
-        assert ret is None
-        return a
-
-
-@testing.parameterize(
-    *testing.product(
-        {'shape': [(0,), (1,), (2, 3), (2, 3, 4)],
-         'values_shape': [(2,), (3, 1), (5,)]}))
-@testing.gpu
-class TestPutmaskNonEqual(unittest.TestCase):
-
-    @testing.for_all_dtypes()
-    @testing.numpy_cupy_allclose()
-    def test_putmask(self, xp, dtype):
-        a = testing.shaped_random(self.shape, xp, dtype=dtype, seed=3)
-        mask = testing.shaped_random(self.shape, xp, dtype=bool, seed=4)
-        values = testing.shaped_random(self.values_shape,
-                                       xp, dtype=dtype, seed=5)
-        ret = xp.putmask(a, mask, values)
-        assert ret is None
-        return a
-
-
-@testing.gpu
-class TestPutmask(unittest.TestCase):
-
-    @testing.numpy_cupy_array_equal()
-    def test_putmask_scalar_values(self, xp):
-        shape = (2, 3)
-        a = testing.shaped_arange(shape, xp)
-        xp.putmask(a, a > 1, 30)
-        return a
-
-    def test_putmask_non_equal_shape_raises(self):
-        for xp in (numpy, cupy):
-            a = xp.array([1, 2, 3])
-            mask = xp.array([True, False])
-            with pytest.raises(ValueError):
-                xp.putmask(a, mask, a**2)
-
-
-class TestPutmaskDifferentDtypes(unittest.TestCase):
-
-    @testing.for_all_dtypes_combination(names=['a_dtype', 'val_dtype'])
-    def test_putmask_differnt_dtypes_raises(self, a_dtype, val_dtype):
-        shape = (2, 3)
-        for xp in (numpy, cupy):
-            a = testing.shaped_random(shape, xp, dtype=a_dtype)
-            mask = testing.shaped_random(shape, xp, dtype=bool)
-            values = testing.shaped_random((3,), xp, dtype=val_dtype)
-            if not numpy.can_cast(val_dtype, a_dtype):
-                with pytest.raises(TypeError):
-                    xp.putmask(a, mask, values)
-
-    @testing.for_all_dtypes()
-    @testing.numpy_cupy_array_equal()
-    def test_putmask_differnt_dtypes_mask(self, xp, dtype):
-        shape = (2, 3)
-        a = testing.shaped_random(shape, xp, dtype=numpy.int64)
-        mask = testing.shaped_random(shape, xp, dtype=dtype)
-        values = testing.shaped_random((3,), xp, dtype=numpy.int64)
-        xp.putmask(a, mask, values)
-        return a
